Add Sensor
Add the sensor abstraction.
